adicionado o webviewpage no ios

+++ memorize/memorize/AspectVGrid.swift
@@ -20,7 +20,7 @@ struct AspectVGrid<Item: Identifiable, ItemView: View>: View {
     
         var body: some View {
             GeometryReader { geometry in
-                let gridItemSize = gridWidthThatFits(count: items.count, size: geometry.size, atAspectRatio: aspectRatio)
+                let gridItemSize = gridWidthThatFits(count: CGFloat(items.count), size: geometry.size, atAspectRatio: aspectRatio)
                 LazyVGrid(columns: [GridItem(.adaptive(minimum: gridItemSize), spacing: 0)], spacing: 0){
                     ForEach(items){ item in
                         content(item)
@@ -32,7 +32,7 @@ struct AspectVGrid<Item: Identifiable, ItemView: View>: View {
     }
     
     func gridWidthThatFits(
-        count: Int,
+        count: CGFloat,
         size: CGSize,
         atAspectRatio aspectRatio: CGFloat
     ) -> CGFloat {
